+++ R/arguments.R
@@ -78,6 +78,13 @@ set_args.model_spec <- function(object, ...) {
   )
 }
 
+#' @export
+set_args.default <- function(object,...) {
+  error_set_object(object, func = "set_args")
+
+  invisible(FALSE)
+}
+
 #' @rdname set_args
 #' @export
 set_mode <- function(object, mode) {
@@ -104,6 +111,13 @@ set_mode.model_spec <- function(object, mode) {
   object$mode <- mode
   object$user_specified_mode <- TRUE
   object
+}
+
+#' @export
+set_mode.default <- function(object, mode) {
+  error_set_object(object, func = "set_mode")
+
+  invisible(FALSE)
 }
 
 # ------------------------------------------------------------------------------